+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 [Changed]
 - Connectors methods are not static for now.
 - Added constructors to Adapters, which allows specified config file for each connections.
+- New properties voId and uniqueName in Group model
  
  ## [v1.0.0]